+++ R/plot.R
@@ -44,14 +44,14 @@ plotValues <- function(data, model, group = "group", raw = "raw") {
 #' Please check the function for inconsistent curves: The different
 #' curves should not intersect. Violations of this assumption are a strong
 #' indication for problems
-#' in modeling the relationship between raw and norm values. There are
+#' in modeling the relationship between raw and norm scores. There are
 #' several reasons, why this might occur:
 #' \enumerate{
-#'   \item Vertical extrapolation: Choosing extreme norm values, e. g. values
-#'   -3 <= x and x >= 3 In order to model these extreme values, a large sample
+#'   \item Vertical extrapolation: Choosing extreme norm scores, e. g. scores
+#'   -3 <= x and x >= 3 In order to model these extreme scores, a large sample
 #'   dataset is necessary.
 #'   \item Horizontal extrapolation: Taylor polynomials converge in a certain
-#'   radius. Using the model values outside the original dataset may
+#'   radius. Using the model scores outside the original dataset may
 #'   lead to inconsistent results.
 #'   \item The data cannot be modeled with Taylor polynomials, or you need
 #'   another power parameter (k) or R2 for the model.
@@ -61,14 +61,14 @@ plotValues <- function(data, model, group = "group", raw = "raw") {
 #' be handled with caution.
 #' checkConsistency and derivationPlot can be used to further inspect the model.
 #' @param model The model from the bestModel function
-#' @param normList Vector with norm values to display
+#' @param normList Vector with norm scores to display
 #' @param minAge Age to start with checking
 #' @param maxAge Upper end of the age check
 #' @param step Stepping parameter for the age check, usually 1 or 0.1; lower
-#' values indicate higher precision / closer checks
-#' @param minRaw Lower end of the raw value range, used for clipping implausible results
+#' scores indicate higher precision / closer checks
+#' @param minRaw Lower end of the raw score range, used for clipping implausible results
 #' (default = 0)
-#' @param maxRaw Upper end of the raw value range, used for clipping implausible results
+#' @param maxRaw Upper end of the raw score range, used for clipping implausible results
 #' @seealso checkConsistency, derivationPlot, plotPercentiles
 #' @examples
 #' # Load example data set, compute model and plot results
@@ -116,7 +116,7 @@ plotNormCurves <- function(model, normList = c(30, 40, 50, 60, 70),
     panel = function(...)
       lattice::panel.superpose(..., panel.groups = panelfun),
     main = "Norm Curves",
-    ylab = "Raw Value", xlab = "Age",
+    ylab = "Raw Score", xlab = "Age",
     col = COL, lwd = 2, grid = TRUE,
     key = list(
       corner = c(0.99, 0.1),
@@ -136,7 +136,7 @@ plotNormCurves <- function(model, normList = c(30, 40, 50, 60, 70),
 #' please check for inconsistent curves, especially intersections.
 #' Violations of this assumption are a strong
 #' indication for problems
-#' in modeling the relationship between raw and norm values.
+#' in modeling the relationship between raw and norm scores.
 #' In general, extrapolation (point 1 and 2) can carefully be done to a
 #' certain degree outside the original sample, but it should in general
 #' be handled with caution.
@@ -145,17 +145,17 @@ plotNormCurves <- function(model, normList = c(30, 40, 50, 60, 70),
 #' Please note, that the estimation of the percentiles of the raw data is done with
 #' the stats::quantile function with the default settings. Please consult help(quantile)
 #' and change the 'type' parameter accordingly.
-#' @param data The raw data including the percentiles and norm values
+#' @param data The raw data including the percentiles and norm scores
 #' @param model The model from the bestModel function
-#' @param minRaw Lower bound of the raw value scale (default = 0)
-#' @param maxRaw Upper bound of the raw value scale
+#' @param minRaw Lower bound of the raw score (default = 0)
+#' @param maxRaw Upper bound of the raw score
 #' @param minAge Variable to restrict the lower bound of the plot to a specific age
 #' @param maxAge Variable to restrict the upper bound of the plot to a specific age
 #' @param raw The name of the raw variable
 #' @param group The name of the grouping variable; the distinct groups are automatically
 #' determined
-#' @param percentiles Vector with percentile values, ranging from 0 to 1 (exclusive)
-#' @param scale The norm value scale, either 'T' (default), 'IQ', 'z', 'percentile' or
+#' @param percentiles Vector with percentile scores, ranging from 0 to 1 (exclusive)
+#' @param scale The norm scale, either 'T' (default), 'IQ', 'z', 'percentile' or
 #' self defined with a double vector with the mean and standard deviation,
 #' f. e. c(10, 3) for Wechsler scale index points
 #' @param type The type parameter of the quantile function to estimate the percentiles
@@ -179,7 +179,7 @@ plotPercentiles <- function(data,
                             scale = "T",
                             type = 7) {
   if (!(raw %in% colnames(data))) {
-    stop(paste(c("ERROR: Raw value variable '", raw, "' does not exist in data object."), collapse = ""))
+    stop(paste(c("ERROR: Raw score variable '", raw, "' does not exist in data object."), collapse = ""))
   }
 
   if (!(group %in% colnames(data))) {
@@ -195,7 +195,7 @@ plotPercentiles <- function(data,
     maxAge <- model$maxA1
   }
 
-  # compute norm values from percentile vector
+  # compute norm scores from percentile vector
   if ((typeof(scale) == "double" && length(scale) == 2)) {
     T <- stats::qnorm(percentiles, scale[1], scale[2])
   } else if (scale == "IQ") {
@@ -268,8 +268,8 @@ plotPercentiles <- function(data,
     i <- i + 1
   }
 
-  # merge actual and predicted values und plot them show lines
-  # for predicted values and dots for actual values
+  # merge actual and predicted scores und plot them show lines
+  # for predicted scores and dots for actual scores
   percentile <- merge(percentile.actual, percentile.fitted,
     by = group, all.y = TRUE
   )
@@ -289,8 +289,8 @@ plotPercentiles <- function(data,
   lattice::xyplot(stats::formula(xyFunction), percentile,
     panel = function(...)
       lattice::panel.superpose(..., panel.groups = panelfun),
-    main = "Actual and predicted percentile curves",
-    ylab = paste0("Raw value (", raw, ")"), xlab = paste0("Explanatory Variable (", group, ")"),
+    main = "Manifest and fitted percentile curves",
+    ylab = paste0("Raw score (", raw, ")"), xlab = paste0("Explanatory Variable (", group, ")"),
     col = COL2, lwd = 2, grid = TRUE,
     key = list(
       corner = c(0.99, 0.01),
@@ -315,19 +315,20 @@ plotPercentiles <- function(data,
 #' negative. The R2 cutoff that was specified in the bestModel function is
 #' displayed as a dashed line.
 #' @param model The regression model from the bestModel function
-#' @param bic Display the log transformed Mallow's Cp (bic = FALSE; default)
-#' or the Bayesian Information Criterion (BIC)
-#' @seealso bestModel, plotPercentiles
+#' @param type Type of chart with 0 = adjusted R2 by number of predictors,
+#' 1 = log transformed Mallow's Cp by adjusted R2 and 2 = Bayesian Information
+#' Criterion (BIC) by adjusted R2
+#' @seealso bestModel, plotPercentiles, printSubset
 #' @examples
 #' normData <- prepareData()
 #' m <- bestModel(data = normData)
 #' plotSubset(m)
 #' @export
-plotSubset <- function(model, bic = FALSE) {
+plotSubset <- function(model, type = 1) {
   message("Hint: Select the model with the highest BIC or Cp score while simultaneously optimizing R2.")
-  if (!bic) {
-    dataFrameTMP <- data.frame(model$subsets$adjr2, model$subsets$cp)
-    lattice::xyplot(log(model.subsets.cp) ~ model.subsets.adjr2,
+  dataFrameTMP <- data.frame(adjr2=model$subsets$adjr2, bic=model$subsets$bic, cp = model$subsets$cp, nr = seq(1, length(model$subsets$adjr2), by=1))
+  if (type == 1) {
+    lattice::xyplot(cp ~ adjr2,
       data = dataFrameTMP, type = "b",
       col.line = "lightblue", lwd = 1,
       grid = TRUE, scales = list(y = list(log = 10)),
@@ -355,9 +356,8 @@ plotSubset <- function(model, bic = FALSE) {
         lattice::panel.xyplot(x, y, ...)
       }
     )
-  } else {
-    dataFrameTMP <- data.frame(model$subsets$adjr2, model$subsets$bic)
-    lattice::xyplot(model.subsets.bic ~ model.subsets.adjr2,
+  } else if(type==2){
+    lattice::xyplot(bic ~ adjr2,
       data = dataFrameTMP, type = "b",
       col.line = "lightblue", lwd = 1,
       grid = TRUE,
@@ -385,33 +385,61 @@ plotSubset <- function(model, bic = FALSE) {
         lattice::panel.xyplot(x, y, ...)
       }
     )
+  }else{
+    lattice::xyplot(adjr2 ~ nr,
+                    data = dataFrameTMP, type = "b",
+                    col.line = "lightblue", lwd = 1,
+                    grid = TRUE,
+                    main = "Information Function",
+                    ylab = "Adjusted R2",
+                    xlab = "Number of predictors",
+                    key = list(
+                      corner = c(
+                        0.9,
+                        0.1
+                      ), lines = list(
+                        col = c("#9933FF"),
+                        lty = c(2), lwd = 2
+                      ),
+                      text = list(c("cutoff value"
+                      ))
+                    ), panel = function(x, y, ...) {
+                      lattice::panel.abline(
+                        h = model$cutoff,
+                        lwd = 2, lty = "longdash",
+                        col = "#9933FF", label = model$cutoff
+                      )
+                      lattice::panel.xyplot(x, y, ...)
+                    }
+    )
+
   }
 }
 
 #' Plot first order derivation of regression model
 #'
 #' Plots the values obtained via the first derivation of the regression model
-#' in dependence of the norm value. The results indicate the progression of the
-#' norm values within each age group. The regression based modeling approach
-#' relies on the assumption of a linear progression of the norm values.
-#' Negative values in the first order derivation indicate a violation of this
-#' assumption. Values near zero
+#' in dependence of the norm score. The results indicate the progression of the
+#' norm scores within each age group. The regression based modeling approach
+#' relies on the assumption of a linear progression of the norm scores.
+#' Negative scores in the first order derivation indicate a violation of this
+#' assumption. Scores near zero
 #' are typical for bottom and ceiling effects in the raw data.
 #' The regression models usually converge within the range of the original
 #' values. In case of vertical and horizontal extrapolation, with increasing
 #' distance to the original data, the risk of assumption violation increases
 #' as well.
 #' ATTENTION: plotDerivative is currently still incompatible with reversed raw
-#' value scales ('descent' option)
+#' score scales ('descent' option)
 #' @param model The model from the bestModel function
 #' @param minAge Age to start with checking
 #' @param maxAge Upper end of the age check
 #' @param stepAge Stepping parameter for the age check, usually 1 or 0.1; lower
 #' values indicate higher precision / closer checks
-#' @param minNorm Lower end of the norm value range, in case of T values, 25 might be good
-#' @param maxNorm Upper end of the norm value range, in case of T values, 25 might be good
-#' @param stepNorm Stepping parameter for norm values
-#' @param descend Reverse raw value order. If set to TRUE, lower raw values
+#' @param minNorm Lower end of the norm score range, in case of T scores, 25 might be good
+#' @param maxNorm Upper end of the norm score range, in case of T scores, 25 might be good
+#' @param stepNorm Stepping parameter for norm scores
+#' @param descend Reverse raw score order. If set to TRUE, lower raw scores
 #' indicate higher performance. Relevant f. e. in case of modelling errors
 #' @seealso checkConsistency, bestModel, derive
 #' @examples
@@ -467,7 +495,7 @@ plotDerivative <- function(model,
       col.regions = regions,
       panel = latticeExtra::panel.2dsmoother,
       main = "Slope of the Regression Function\n(1st order derivation)",
-      ylab = "1st order derivate of norm value",
+      ylab = "1st order derivate of norm score",
       xlab = "Age"
     )
   } else {
@@ -477,7 +505,7 @@ plotDerivative <- function(model,
       colorkey = key,
       col.regions = regions,
       main = "Slope of the Regression Function\n(1st order derivation)",
-      ylab = "1st order derivate of norm value",
+      ylab = "1st order derivate of norm score",
       xlab = "Age"
     )
   }